+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 
 * Add support for saving Azure key version with DEK (#2641)
 * Pass context when clients make KEK calls to DEK Registry (#2642)
+* Add CI coverage for linux-s390x (IBM Z) and document s390x platform support
 
 ## Fixes
 
 * Fix early return in FindMessageByName (#2644)
+* Fix the CSFLE ciphertext version field being written in host byte order on big-endian
+  platforms, which made encrypted fields unreadable across architectures
 
 
 # 2.15.0